refactor: move ingressNeedsUpdate tests to the unit test file
