+++ ray-operator/controllers/ray/raycluster_controller_unit_test.go
@@ -4298,3 +4298,165 @@ func TestReconcileIngressKubernetesSkipsUnownedIngressWithGeneratedName(t *testi
 		t.Fatal("expected a warning event for the ingress name collision, but none was recorded")
 	}
 }
+
+func TestIngressNeedsUpdateAppliesDesiredFields(t *testing.T) {
+	exact := networkingv1.PathTypeExact
+	prefix := networkingv1.PathTypePrefix
+
+	existing := &networkingv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "sample-head-ingress",
+			Namespace: "default",
+			Labels: map[string]string{
+				"ray.io/cluster": "sample",
+			},
+			Annotations: map[string]string{
+				"nginx.ingress.kubernetes.io/rewrite-target": "/$1",
+			},
+		},
+		Spec: networkingv1.IngressSpec{
+			Rules: []networkingv1.IngressRule{
+				{
+					Host: "old.example.com",
+					IngressRuleValue: networkingv1.IngressRuleValue{
+						HTTP: &networkingv1.HTTPIngressRuleValue{Paths: []networkingv1.HTTPIngressPath{{Path: "/sample/(.*)", PathType: &exact}}},
+					},
+				},
+			},
+		},
+	}
+
+	desired := &networkingv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "sample-head-ingress",
+			Namespace: "default",
+			Labels: map[string]string{
+				"ray.io/cluster": "sample",
+				"extra":          "label",
+			},
+			Annotations: map[string]string{
+				"nginx.ingress.kubernetes.io/ssl-redirect": "true",
+			},
+		},
+		Spec: networkingv1.IngressSpec{
+			Rules: []networkingv1.IngressRule{
+				{
+					Host: "new.example.com",
+					IngressRuleValue: networkingv1.IngressRuleValue{
+						HTTP: &networkingv1.HTTPIngressRuleValue{Paths: []networkingv1.HTTPIngressPath{{Path: "/dashboard", PathType: &prefix}}},
+					},
+				},
+			},
+			TLS: []networkingv1.IngressTLS{{Hosts: []string{"new.example.com"}, SecretName: "ray-tls"}},
+		},
+	}
+
+	updated := ingressNeedsUpdate(existing, desired)
+
+	assert.True(t, updated)
+	assert.Equal(t, desired.Labels, existing.Labels)
+	assert.Equal(t, desired.Annotations, existing.Annotations)
+	assert.Equal(t, desired.Spec, existing.Spec)
+}
+
+func TestIngressNeedsUpdateNoopWhenEqual(t *testing.T) {
+	prefix := networkingv1.PathTypePrefix
+	existing := &networkingv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "sample-head-ingress",
+			Namespace: "default",
+			Labels: map[string]string{
+				"ray.io/cluster": "sample",
+			},
+			Annotations: map[string]string{
+				"nginx.ingress.kubernetes.io/ssl-redirect": "true",
+			},
+		},
+		Spec: networkingv1.IngressSpec{
+			Rules: []networkingv1.IngressRule{
+				{
+					Host: "new.example.com",
+					IngressRuleValue: networkingv1.IngressRuleValue{
+						HTTP: &networkingv1.HTTPIngressRuleValue{Paths: []networkingv1.HTTPIngressPath{{Path: "/dashboard", PathType: &prefix}}},
+					},
+				},
+			},
+			TLS: []networkingv1.IngressTLS{{Hosts: []string{"new.example.com"}, SecretName: "ray-tls"}},
+		},
+	}
+
+	desired := existing.DeepCopy()
+
+	updated := ingressNeedsUpdate(existing, desired)
+
+	assert.False(t, updated)
+}
+
+func TestIngressNeedsUpdatePreservesDefaultedIngressClassName(t *testing.T) {
+	prefix := networkingv1.PathTypePrefix
+	rules := []networkingv1.IngressRule{
+		{
+			Host: "ray.example.com",
+			IngressRuleValue: networkingv1.IngressRuleValue{
+				HTTP: &networkingv1.HTTPIngressRuleValue{Paths: []networkingv1.HTTPIngressPath{{Path: "/", PathType: &prefix}}},
+			},
+		},
+	}
+
+	// The live Ingress has spec.ingressClassName set by the DefaultIngressClass
+	// admission plugin, while the desired object (the RayCluster did not request a
+	// class) leaves it nil.
+	existing := &networkingv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{Name: "sample-head-ingress", Namespace: "default"},
+		Spec: networkingv1.IngressSpec{
+			IngressClassName: ptr.To("nginx"),
+			Rules:            rules,
+		},
+	}
+	desired := &networkingv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{Name: "sample-head-ingress", Namespace: "default"},
+		Spec: networkingv1.IngressSpec{
+			Rules: rules,
+		},
+	}
+
+	updated := ingressNeedsUpdate(existing, desired)
+
+	assert.False(t, updated, "should not update when the only difference is an admission-defaulted ingress class")
+	require.NotNil(t, existing.Spec.IngressClassName)
+	assert.Equal(t, "nginx", *existing.Spec.IngressClassName, "the defaulted ingress class must be preserved")
+}
+
+func TestIngressNeedsUpdateOverridesIngressClassNameWhenRequested(t *testing.T) {
+	prefix := networkingv1.PathTypePrefix
+	rules := []networkingv1.IngressRule{
+		{
+			Host: "ray.example.com",
+			IngressRuleValue: networkingv1.IngressRuleValue{
+				HTTP: &networkingv1.HTTPIngressRuleValue{Paths: []networkingv1.HTTPIngressPath{{Path: "/", PathType: &prefix}}},
+			},
+		},
+	}
+
+	// When the RayCluster does request a class, it must win over the live value.
+	existing := &networkingv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{Name: "sample-head-ingress", Namespace: "default"},
+		Spec: networkingv1.IngressSpec{
+			IngressClassName: ptr.To("nginx"),
+			Rules:            rules,
+		},
+	}
+	desired := &networkingv1.Ingress{
+		ObjectMeta: metav1.ObjectMeta{Name: "sample-head-ingress", Namespace: "default"},
+		Spec: networkingv1.IngressSpec{
+			IngressClassName: ptr.To("traefik"),
+			Rules:            rules,
+		},
+	}
+
+	updated := ingressNeedsUpdate(existing, desired)
+
+	assert.True(t, updated)
+	require.NotNil(t, existing.Spec.IngressClassName)
+	assert.Equal(t, "traefik", *existing.Spec.IngressClassName, "an explicitly requested ingress class must override the live value")
+}
